+++ src/main.cpp
@@ -5,13 +5,38 @@
 
 using namespace geode::prelude;
 
-std::string createXLString(int levelLengthMinutes) {
-	std::string XLlabel;
-	for (int i = 0; i < (int) log2(levelLengthMinutes) & i < 10; i++) {
-		XLlabel = XLlabel.append("X");
+std::string createXLstring(int levelLengthMinutes) {
+	std::string XLstring;
+	int maximumXs = Mod::get()->getSettingValue<int64_t>("maximum-xs");
+	bool usingPowerNotation = Mod::get()->getSettingValue<bool>("use-power-notation");
+	bool usingXXLplus = Mod::get()->getSettingValue<bool>("xxl-plus");
+	int lengthExponent = log2(levelLengthMinutes);
+
+	if (usingPowerNotation && lengthExponent > 1) {
+		XLstring.append("X^");
+		XLstring.append(std::to_string(lengthExponent));
 	}
-	XLlabel = XLlabel.append("L");
-	return XLlabel;
+
+	else {
+		if (lengthExponent > maximumXs) {
+			lengthExponent = maximumXs;
+		}
+		for (int i = 0; i < lengthExponent; i++) {
+			XLstring.append("X");
+		}
+	}
+	XLstring.append("L");
+
+	if (!usingXXLplus) {
+		return XLstring;
+	}
+
+    int XXLPlusLength = pow(2, lengthExponent) + pow(2, lengthExponent - 1);
+    if (levelLengthMinutes >= XXLPlusLength) {
+		XLstring.append("+");
+    }
+
+    return XLstring;
 };
 
 class $modify(MyLevelInfoLayer, LevelInfoLayer) {
@@ -24,8 +49,8 @@ class $modify(MyLevelInfoLayer, LevelInfoLayer) {
 			int cvoltonLengthMinutes = timeForLevelString(m_level->m_levelString);
 
 			Loader::get()->queueInMainThread([this,cvoltonLengthMinutes]() {
-				if (cvoltonLengthMinutes >= 4) {
-					m_lengthLabel->setString(createXLString(cvoltonLengthMinutes).c_str());
+				if (cvoltonLengthMinutes >= 3) {
+					m_lengthLabel->setString(createXLstring(cvoltonLengthMinutes).c_str());
 				}
 				this->release();
 			});
@@ -45,11 +70,11 @@ class $modify(MyLevelInfoLayer, LevelInfoLayer) {
 			return;
 		}
 
-		if (levelLengthMinutes < 4) {
+		if (levelLengthMinutes < 3) {
 			return;
 		}
 
-		m_lengthLabel->setString(createXLString(levelLengthMinutes).c_str());
+		m_lengthLabel->setString(createXLstring(levelLengthMinutes).c_str());
 	}
 
 	void setupLevelInfo() {